+++ packages/openapi-typegen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sebspark/openapi-typegen
 
+## 1.11.0
+
+### Minor Changes
+
+- 37984eb: Enums are now exported as a const (MY_ENUM_VALUES) and a type (MyEnum)
+
+## 1.10.0
+
+### Minor Changes
+
+- 894bf2e: remove yarn as a dependency
+
 ## 1.9.0
 
 ### Minor Changes