+++ packages/promise-cache/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sebspark/promise-cache
 
+## 2.1.1
+
+### Patch Changes
+
+- df6468d: ttl in response should be in seconds
+
+## 2.1.0
+
+### Minor Changes
+
+- 6d40aae: enable setting cache ttl from promise response
+
 ## 2.0.10
 
 ### Patch Changes